# api.yaml
openapi: 3.1.0
...
info:
  title: Chatbot Administrasi Mahasiswa
  version: 0.1.0
  description: |
    API chatbot RAG untuk pertanyaan administrasi akademik mahasiswa.

    ## Lima jenis balasan

    Endpoint chat selalu mengembalikan 200 dengan `kind` yang menentukan cara
    frontend menampilkannya. `answer`, `refusal`, dan `support` WAJIB terlihat
    berbeda satu sama lain:

    | `kind` | Arti | Tampilan |
    |---|---|---|
    | `answer` | Jawaban bersumber dokumen resmi | Bubble normal + kartu sitasi (FE-2) |
    | `refusal` | Retrieval terlalu lemah, LLM tidak dipanggil (FR-3) | Visual berbeda + kontak (FE-4) |
    | `support` | Pertanyaan bernuansa tekanan mental (FR-7) | Nada empatik + kontak konseling |
    | `smalltalk` | Sapaan atau basa-basi; retrieval dan LLM dilewati | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |
    | `rejected` | Dihentikan gerbang JEV: nonsense, upaya manipulasi, atau di luar topik kampus | Bubble biasa, tanpa sitasi dan tanpa tombol umpan balik |

    Hanya `answer` yang membawa sitasi. Menampilkan kartu sitasi di layar
    penolakan memberi kesan keliru bahwa jawabannya bersumber; menampilkannya
    pada sapaan sama saja mengklaim "halo" ada di panduan akademik.

    `smalltalk` ada supaya mahasiswa yang mengetik "hai" tidak dibalas
    penolakan berisi alamat loket biro, dan supaya sapaan tidak masuk daftar
    pertanyaan tak terjawab (AD-4).

    `rejected` hanya muncul bila gerbang JEV dinyalakan (`JEV_ENABLED=true`).
...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/chat:
    post:
      tags: [chat]
      summary: Ajukan pertanyaan, terima jawaban lengkap
      operationId: chat
      x-status: implemented
      description: |
        Balasan sekali kirim. Dipakai kotak uji coba admin (AD-6) dan test.
        Untuk antarmuka mahasiswa gunakan `/api/chat/stream` agar token muncul
        bertahap (FE-1).

        Alurnya (PRD §7): sanitasi input -> pemeriksaan sensitif (FR-7) ->
        penulisan ulang query (FR-4) -> retrieval hibrida (FR-2) -> ambang
        penolakan (FR-3) -> deteksi topik berisiko (FR-6) -> generasi (FR-5).

        LLM tidak dipanggil sama sekali bila `kind` bernilai `refusal` atau
        `support`.
      parameters:
        - $ref: '#/components/parameters/SessionIdHeader'
        - $ref: '#/components/parameters/EmbedKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              first_question:
                summary: Pesan pertama, tanpa riwayat
                value:
                  question: Kapan pengisian KRS semester ganjil dibuka?
                  session_id: 5f2c9a1b-7e04-4d3a-9c88-1a2b3c4d5e6f
              follow_up_question:
                summary: Pertanyaan lanjutan, memicu penulisan ulang query (FR-4)
                value:
                  question: Kalau telat gimana?
                  session_id: 5f2c9a1b-7e04-4d3a-9c88-1a2b3c4d5e6f
                  history:
                    - role: user
                      content: Kapan pengisian KRS dibuka?
                    - role: assistant
                      content: Pengisian KRS dibuka 1-7 Agustus 2025.
              with_unit:
                summary: Dibatasi pada unit yang dipilih mahasiswa di menu
                value:
                  question: Bagaimana cara membayar UKT lewat virtual account?
                  session_id: 5f2c9a1b-7e04-4d3a-9c88-1a2b3c4d5e6f
                  unit: Keuangan
      responses:
        '200':
          description: |
            Selalu 200 selama pertanyaannya sah, termasuk saat sistem menolak
            menjawab. Penolakan adalah hasil yang sah, bukan kegagalan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                answer:
                  $ref: '#/components/examples/NormalAnswer'
                refusal:
                  $ref: '#/components/examples/Refusal'
                escalation:
                  $ref: '#/components/examples/EscalatedAnswer'
                support:
                  $ref: '#/components/examples/Support'
                smalltalk:
                  $ref: '#/components/examples/Smalltalk'
        '403':
          $ref: '#/components/responses/InvalidEmbedKey'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/KillSwitchEngaged'
...
  /api/feedback:
    post:
      tags: [chat]
      summary: Kirim umpan balik atas satu jawaban
      operationId: submit_feedback
      x-status: implemented
      description: |
        FE-5: satu klik, tanpa modal. `comment` opsional dan hanya muncul bila
        mahasiswa memilih mengisinya.

        Mengirim ulang untuk `message_id` yang sama MENGGANTI umpan balik
        sebelumnya, supaya mahasiswa yang berubah pikiran dari 👍 ke 👎 tidak
        terhitung dua kali.

        Rasio umpan balik positif adalah metrik PRD §3 (target >= 75%).
        Perhatikan anti-metriknya: volume percakapan tinggi bisa berarti sistem
        berguna, bisa juga berarti jawabannya buruk sehingga mahasiswa bertanya
        berulang kali.
      parameters:
        - $ref: '#/components/parameters/SessionIdHeader'
        - $ref: '#/components/parameters/EmbedKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '204':
          description: Tercatat. Tidak ada isi balasan -- FE-5 tidak menampilkan konfirmasi.
        '403':
          $ref: '#/components/responses/InvalidEmbedKey'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /api/admin/unanswered:
    get:
      tags: [admin-quality]
      summary: Pertanyaan tak terjawab, dikelompokkan
      operationId: list_unanswered
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        AD-4. PRD menyebut halaman ini "menentukan sistem membaik atau stagnan"
        dan meminta ia ditempatkan menonjol di navigasi.

        Pertanyaan yang mirip dikelompokkan beserta frekuensinya, sehingga admin
        melihat "12 mahasiswa menanyakan ini" alih-alih 12 baris terpisah.

        Kemiripan diukur dari kata yang sama setelah kata tanya dan sapaan
        dibuang (bukan embedding), jadi dapat dijelaskan dan tidak memanggil API
        berbayar. Pertanyaan yang sudah dan belum ditindaklanjuti tidak pernah
        berada di kelompok yang sama. `limit` dan `offset` berlaku atas
        kelompok, bukan atas baris.
      parameters:
        - name: resolved
          in: query
          description: Kosongkan untuk menampilkan keduanya.
          schema: { type: boolean }
        - name: since
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Kelompok pertanyaan
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UnansweredGroup' }
              example:
                - ids:
                    - 0b6f2c1e-3a4d-4e5f-9a8b-7c6d5e4f3a2b
                    - 1c7a3d2f-4b5e-4f60-8b9c-8d7e6f5a4b3c
                  sample_question: Bagaimana cara mengurus surat keterangan lulus?
                  count: 12
                  avg_top_score: 0.21
                  last_asked_at: '2026-09-09T14:22:11Z'
                  resolved: false
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/unanswered/{unanswered_id}:
    patch:
      tags: [admin-quality]
      summary: Tandai pertanyaan sudah ditindaklanjuti
      operationId: resolve_unanswered
      x-status: implemented
      x-min-role: admin
      security: [{ bearerAuth: [] }]
      description: |
        Ditandai `resolved` setelah dokumen yang menjawabnya diunggah. Ini
        bukan sekadar penanda kerapian -- selisih antara pertanyaan masuk dan
        pertanyaan terselesaikan adalah ukuran apakah sistem benar-benar
        membaik.

        Satu PATCH mengubah satu baris. Untuk menandai satu kelompok AD-4,
        kirim PATCH untuk setiap id di `UnansweredGroup.ids`.
      parameters:
        - name: unanswered_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnansweredUpdate' }
      responses:
        '204': { description: Tersimpan }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/admin/feedback:
    get:
      tags: [admin-quality]
      summary: Umpan balik mahasiswa atas jawaban
      operationId: list_feedback
      x-status: implemented
      x-min-role: admin
      security: [{ bearerAuth: [] }]
      description: |
        Penilaian 👍/👎 yang dikirim mahasiswa lewat `POST /api/feedback` (FE-5),
        terbaru lebih dulu, beserta pertanyaan dan jawaban yang dinilai.

        Statistik AD-5 hanya memuat rasio kepuasan: ia memberi tahu ada yang
        salah, bukan apanya. Daftar ini melengkapinya -- jempol ke bawah dapat
        ditelusuri ke pertanyaan yang memicunya, lalu diuji ulang lewat
        `POST /api/admin/test-query`.

        `comment` jarang terisi: FE-5 satu klik tanpa modal, jadi sebagian besar
        baris hanya berupa jempol. `question` diambil dari pesan mahasiswa
        terakhir sebelum jawaban itu di percakapan yang sama; untuk pertanyaan
        sensitif (FR-7) yang tersimpan adalah penanda tetap, bukan kalimat
        aslinya.

        Isi percakapan, jadi levelnya sama dengan statistik: minimal admin.
      parameters:
        - name: helpful
          in: query
          description: Kosongkan untuk menampilkan keduanya.
          schema: { type: boolean }
        - name: since
          in: query
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Umpan balik
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FeedbackPage' }
              example:
                items:
                  - id: 7d1e4b2a-9c3f-4e58-b0a7-1f2e3d4c5b6a
                    message_id: 9c3e1a44-6b2d-4f51-8a70-2d9b5c1e7f03
                    helpful: false
                    comment: Jawabannya tidak menyebut biaya cetaknya
                    created_at: '2026-09-09T14:25:03Z'
                    question: Berapa biaya legalisir ijazah?
                    answer: Legalisir ijazah diurus di Biro Administrasi Akademik [Panduan Akademik 2025, hal. 12].
                    kind: answer
                    top_score: 0.62
                total: 18
                positive_count: 124
                negative_count: 18
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/admin/test-query:
    post:
      tags: [admin-quality]
      summary: Uji coba retrieval, tampilkan chunk dan skornya
      operationId: admin_test_query
      x-status: implemented
      x-min-role: staf
      security: [{ bearerAuth: [] }]
      description: |
        AD-6. Berbeda dari `/api/chat`: balasan ikut membawa chunk yang terambil
        beserta skor mentah tiap sumber (vector dan fulltext) dan skor RRF.

        Ini alat diagnosa utama saat ada laporan "jawabannya salah". Skor mentah
        yang ditampilkan, bukan skor RRF, karena skor RRF hanya mencerminkan
        peringkat -- chunk terbaik dari sekumpulan chunk tak relevan tetap
        mendapat skor RRF tertinggi.

        Tidak tunduk pada kill switch (admin perlu mendiagnosis justru saat
        layanan dimatikan) dan TIDAK dicatat ke log percakapan, supaya uji coba
        tidak mencemari AD-4 maupun AD-5. LLM tetap dipanggil bila lolos ambang.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestQueryRequest' }
      responses:
        '200':
          description: Hasil beserta rincian retrieval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestQueryResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token dari `POST /api/admin/login`. Hanya untuk endpoint admin.
  parameters:
    SessionIdHeader:
      name: X-Session-Id
      in: header
      required: false
      description: |
        Pembatasan laju per sesi (FR-9). Bila tidak dikirim, sistem jatuh ke
        pembatasan per IP -- yang di jaringan kampus dipakai bersama ratusan
        mahasiswa, sehingga jauh lebih mudah kena batas. Kirimkan nilai yang
        sama dengan `session_id` di body.
      schema: { type: string, minLength: 8, maxLength: 128 }
    EmbedKeyHeader:
      name: X-Embed-Key
      in: header
      required: false
      description: |
        Kunci situs penyemat, hanya dikirim panel `/embed` portal yang dimuat
        situs lain. Tanpa header = permintaan dari portal sendiri. Kunci yang
        tidak dikenal atau sudah dinonaktifkan dijawab 403.
      schema: { type: string, maxLength: 64 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
  responses:
    ValidationError:
      description: Body permintaan tidak sah
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
    RateLimited:
      description: |
        Melebihi batas laju (FR-9): per IP, per `X-Session-Id`, atau per kunci
        sematan -- mana pun yang lebih dulu penuh (`RATE_LIMIT_*` di `.env`).
        Permintaan yang ditolak tidak ikut menghabiskan jatah. Menilai jawaban
        (`/api/feedback`) punya jatah sendiri, terpisah dari bertanya.
      headers:
        Retry-After:
          description: Detik sampai boleh mencoba lagi.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            detail: Terlalu banyak pertanyaan dalam waktu singkat. Coba lagi sebentar lagi.
    KillSwitchEngaged:
      description: |
        Layanan chat dimatikan operator (FR-9), atau otomatis karena pertanyaan
        hari ini melewati `CHAT_DAILY_LIMIT`. Frontend menampilkan `detail` apa
        adanya -- kalimatnya sudah memuat kontak manusia sebagai jalan keluar.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            detail: >-
              Layanan chat sedang dinonaktifkan sementara. Silakan hubungi Front
              Office INSTIKI: Telepon (0361) 256995 / WhatsApp 0813-3896-9832.
    Unauthorized:
      description: |
        Token tidak ada, tidak valid, atau kedaluwarsa; akun sudah dinonaktifkan
        atau dihapus; atau kata sandi sudah diganti sejak token terbit. Frontend
        menghapus sesi dan kembali ke halaman login.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        Level akun di bawah `x-min-role` operasi ini, atau staf/dosen menyentuh
        dokumen unit lain. Sesi tetap berlaku; `detail` siap tampil.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InvalidEmbedKey:
      description: |
        `X-Embed-Key` dikirim tetapi kuncinya tidak dikenal atau sudah
        dinonaktifkan dari dashboard. Frontend menampilkan `detail` apa adanya.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            detail: >-
              Asisten tidak tersedia di situs ini. Silakan hubungi pengelola situs,
              atau ajukan pertanyaan lewat portal akademik.
  examples:
    NormalAnswer:
      summary: Jawaban bersumber dokumen
      value:
        kind: answer
        message_id: 9c3e1a44-6b2d-4f51-8a70-2d9b5c1e7f03
        text: >-
          Pengisian KRS semester ganjil dibuka pada 1-7 Agustus 2025
          [Panduan Akademik 2025, hal. 12].
        citations:
          - title: Panduan Akademik 2025
            page: 12
            document_id: 3f1a2b3c-4d5e-6f70-8192-a3b4c5d6e7f8
            file_path: storage/documents/3f1a2b3c-4d5e-6f70-8192-a3b4c5d6e7f8.pdf
        contacts: []
        escalated: false
        top_score: 0.82
    EscalatedAnswer:
      summary: Topik berisiko tinggi -- jawaban plus banner kontak (FR-6, FE-3)
      value:
        kind: answer
        message_id: 1b7d9e02-3c4f-4a56-9d81-5e6f7a8b9c0d
        text: >-
          Batas akhir pembayaran UKT semester ganjil adalah 25 Agustus 2025
          [Panduan Akademik 2025, hal. 30].
        citations:
          - title: Panduan Akademik 2025
            page: 30
            document_id: 3f1a2b3c-4d5e-6f70-8192-a3b4c5d6e7f8
            file_path: storage/documents/3f1a2b3c-4d5e-6f70-8192-a3b4c5d6e7f8.pdf
        contacts:
          - unit: Front Office INSTIKI
            service_hours: Senin-Sabtu, 08.00-22.00 WITA
            contact: Telepon (0361) 256995 / WhatsApp 0813-3896-9832
        escalated: true
        top_score: 0.77
    Refusal:
      summary: Retrieval terlalu lemah -- LLM tidak dipanggil (FR-3)
      value:
        kind: refusal
        message_id: 5a2c8f13-7e94-4b60-a1d2-3f4e5a6b7c8d
        text: |-
          Maaf, saya tidak menemukan informasi ini di dokumen resmi yang saya miliki. Supaya Anda tidak mendapat jawaban yang keliru, silakan tanyakan langsung ke:

          - Front Office INSTIKI (Senin-Sabtu, 08.00-22.00 WITA): Telepon (0361) 256995 / WhatsApp 0813-3896-9832
        citations: []
        contacts:
          - unit: Front Office INSTIKI
            service_hours: Senin-Sabtu, 08.00-22.00 WITA
            contact: Telepon (0361) 256995 / WhatsApp 0813-3896-9832
        escalated: true
        top_score: 0.11
    Support:
      summary: Pertanyaan bernuansa tekanan mental (FR-7)
      value:
        kind: support
        message_id: 7e1f4a90-2b35-4c68-9d0a-1b2c3d4e5f60
        text: |-
          Terima kasih sudah menyampaikan ini. Yang Anda rasakan wajar dan Anda tidak sendirian. Saya tidak bisa membantu urusan seperti ini lewat dokumen administrasi, tetapi ada orang yang siap mendengarkan:

          - Unit Layanan Konseling INSTIKI (Jadwal konseling diatur melalui Front Office): Telepon (0361) 256995 / WhatsApp 0813-3896-9832
          - Dosen Wali (Sesuai jadwal bimbingan): Hubungi melalui program studi
        citations: []
        contacts:
          - unit: Unit Layanan Konseling INSTIKI
            service_hours: Jadwal konseling diatur melalui Front Office
            contact: Telepon (0361) 256995 / WhatsApp 0813-3896-9832
          - unit: Dosen Wali
            service_hours: Sesuai jadwal bimbingan
            contact: Hubungi melalui program studi
        escalated: true
        top_score: null
    Smalltalk:
      summary: Sapaan atau basa-basi -- tanpa retrieval dan tanpa LLM
      value:
        kind: smalltalk
        message_id: 2d4f6a81-9c03-4e57-b1a2-6f7e8d9c0b1a
        text: >-
          Halo! Saya asisten administrasi akademik. Silakan tanyakan urusan
          administrasi Anda -- jawaban saya bersumber dari dokumen resmi kampus
          dan selalu menyertakan sumbernya.
        citations: []
        contacts: []
        escalated: false
        top_score: null
  schemas:
...
    ChatRequest:
      type: object
      required: [question, session_id]
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 500
          description: |
            Pertanyaan apa adanya. Server yang menyanitasi dan membungkusnya
            dengan delimiter; klien tidak perlu (dan tidak boleh) ikut memproses.

            500 = panjang maksimum pertanyaan siap klik (entri tanya jawab admin).
            Kotak pertanyaan portal sendiri membatasi 200.
        session_id:
          type: string
          minLength: 8
          maxLength: 128
          description: |
            Dibangkitkan klien, disimpan di localStorage. JANGAN isi dengan NIM,
            email, atau apa pun yang dapat mengenali orang (PRD §11).
        history:
          type: array
          default: []
          maxItems: 50
          description: |
            Riwayat percakapan. Hanya 3 pesan terakhir yang dipakai untuk
            penulisan ulang query (FR-4); sisanya dibuang server, dan lebih dari
            50 ditolak 422.
          items: { $ref: '#/components/schemas/TurnIn' }
        unit:
          type: [string, 'null']
          maxLength: 200
          description: |
            `name` dari `GET /api/units`. Retrieval -- vektor maupun fulltext --
            hanya mencari di dokumen unit ini. Kosongkan atau kirim null untuk
            mencari di semua unit.

            Dicocokkan tanpa peduli huruf besar dan spasi berlebih. Nama yang
            tidak terdaftar, atau unit yang sudah dinonaktifkan, dijawab 422;
            muat ulang daftar unit bila itu terjadi.

            Bila tidak ada dokumen unit ini yang cukup relevan, balasannya
            penolakan biasa (`kind: refusal`) yang menyebut unit tersebut dan
            menyarankan memilih unit lain atau semua unit.
...
    ChatResponse:
      type: object
      required: [kind, text, citations, contacts, escalated]
      properties:
        kind: { $ref: '#/components/schemas/OutcomeKind' }
        message_id:
          type: [string, 'null']
          format: uuid
          description: |
            Id jawaban di tabel `messages`, dipakai `POST /api/feedback`. Bernilai
            null bila pencatatan ke database gagal (jawaban tetap terkirim) --
            frontend harus menyembunyikan tombol 👍/👎 saat null, bukan mengirim
            id kosong.
        text: { type: string }
        citations:
          type: array
          default: []
          description: |
            Hanya dokumen yang benar-benar dikutip di `text`, urut kemunculan --
            bukan seluruh chunk yang terambil retrieval. Retrieval mengambil
            beberapa chunk sebagai konteks LLM; menampilkan semuanya sebagai
            "Sumber" memaksa mahasiswa menebak kartu mana yang relevan.

            Sitasi ke dokumen di luar konteks (sumber karangan) tidak pernah
            ikut: tautannya buntu. Jawaban tanpa satu pun penanda sitasi jatuh
            kembali ke seluruh chunk terambil, supaya verifikasi tetap mungkin.

            Selalu kosong bila `kind` bukan `answer`.
          items: { $ref: '#/components/schemas/CitationOut' }
        contacts:
          type: array
          default: []
          items: { $ref: '#/components/schemas/ContactOut' }
        escalated:
          type: boolean
          default: false
          description: Tampilkan banner FE-3 bila true.
        top_score:
          type: [number, 'null']
          description: |
            Skor kemiripan mentah tertinggi, bukan skor RRF. Jangan ditampilkan
            ke mahasiswa -- angka ini untuk AD-6 dan kalibrasi ambang.
    FeedbackRequest:
      type: object
      required: [message_id, helpful]
      properties:
        message_id: { type: string, format: uuid }
        helpful: { type: boolean, description: 'true = 👍, false = 👎' }
        comment: { type: [string, 'null'], maxLength: 1000 }
...
    UnansweredUpdate:
      type: object
      required: [resolved]
      properties:
        resolved: { type: boolean }
    UnansweredGroup:
      type: object
      required: [ids, sample_question, count, last_asked_at, resolved]
      properties:
        ids:
          type: array
          minItems: 1
          items: { type: string, format: uuid }
          description: |
            Seluruh baris `unanswered_questions` dalam kelompok. Menandai kelompok selesai
            berarti PATCH `/api/admin/unanswered/{id}` untuk setiap id.
        sample_question:
          type: string
          description: Satu pertanyaan mewakili kelompok, apa adanya seperti diketik.
        count: { type: integer, minimum: 1 }
        avg_top_score:
          type: [number, 'null']
          description: |
            Rata-rata skor tertinggi kelompok ini. Nilai yang mendekati ambang
            menandakan dokumennya mungkin sudah ada tetapi retrieval-nya meleset;
            nilai sangat rendah menandakan dokumennya memang belum ada.
        last_asked_at: { type: string, format: date-time }
        resolved: { type: boolean }
...
    FeedbackPage:
      type: object
      required: [items, total, positive_count, negative_count]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/FeedbackItem' }
        total:
          type: integer
          minimum: 0
          description: Jumlah baris yang cocok dengan seluruh filter, untuk penomoran halaman.
        positive_count: { type: integer, minimum: 0 }
        negative_count:
          type: integer
          minimum: 0
          description: |
            Keduanya dihitung mengabaikan filter `helpful`, sehingga jumlah pada
            kedua tab tetap terlihat saat salah satunya sedang dipilih.
...
    TestQueryRequest:
      type: object
      required: [question]
      properties:
        question: { type: string, minLength: 1, maxLength: 2000 }
        vector_threshold:
          description: |
            Timpa ambang vektor untuk uji coba ini saja, untuk mencari nilai yang
            pas. Tidak mengubah ambang yang berlaku bagi mahasiswa.
          type: [number, 'null']
          minimum: 0
          maximum: 1
        unit:
          type: [string, 'null']
          maxLength: 200
          description: |
            Sama dengan `ChatRequest.unit`: uji apa yang dilihat mahasiswa yang
            memilih unit ini di menu chatbot.
...
    TestQueryResponse:
      type: object
      required: [kind, text, retrieved, decision, thresholds, contacts, escalated, latency_ms]
      properties:
        kind: { $ref: '#/components/schemas/OutcomeKind' }
        text: { type: string }
        rewritten_query: { type: [string, 'null'] }
        retrieved:
          type: array
          description: Tetap terisi saat ditolak -- justru itu yang perlu dilihat saat diagnosa.
          items: { $ref: '#/components/schemas/RetrievedChunk' }
        decision:
          type: [object, 'null']
          description: null bila pertanyaan dialihkan ke konseling (FR-7) sebelum retrieval.
          required: [decision, reason]
          properties:
            decision: { type: string, enum: [proceed, refuse] }
            reason: { type: string, enum: [ok, no_results, below_threshold] }
            top_vector_score: { type: [number, 'null'] }
            top_lexical_score: { type: [number, 'null'] }
        thresholds: { $ref: '#/components/schemas/ThresholdValues' }
        contacts:
          type: array
          description: Kontak yang akan dilihat mahasiswa untuk pertanyaan yang sama (FE-3, FE-4).
          items: { $ref: '#/components/schemas/ContactOut' }
        escalated: { type: boolean }
        latency_ms: { type: integer, minimum: 0 }
        langsmith_run_id:
          type: [string, 'null']
          description: >-
            Akar trace LangSmith untuk uji coba ini -- mencakup penulisan ulang query,
            retrieval, dan penyusunan jawaban sekaligus. Null bila tracing mati
            (LANGSMITH_TRACING=false atau LANGSMITH_API_KEY kosong).
...